3-e2e-health-advisor-sample/backend/requirements.txt
aniso8601==10.0.0
annotated-types==0.7.0
anyio==4.8.0
argcomplete==3.5.3
attrs==24.3.0
azure-ai-evaluation==1.1.0
azure-ai-inference==1.0.0b7
azure-ai-projects==1.0.0b5
azure-common==1.1.28
azure-core==1.32.0
azure-identity==1.19.0
azure-monitor-opentelemetry-exporter==1.0.0b33
azure-search-documents==11.5.2
azure-storage-blob==12.24.0
blinker==1.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.0
Deprecated==1.2.15
distro==1.9.0
docstring_parser==0.16
fastapi==0.115.6
filelock==3.16.1
filetype==1.2.0
fixedint==0.1.6
Flask==3.1.0
Flask-Cors==5.0.0
flask-restx==1.3.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.5.2
isodate==0.7.2
itsdangerous==2.2.0
jaraco.classes==3.4.0
jeepney==0.8.0
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
keyring==24.3.1
MarkupSafe==3.0.2
marshmallow==3.25.1
more-itertools==10.6.0
msal==1.31.1
msal-extensions==1.2.0
msrest==0.7.1
nltk==3.9.1
numpy==2.2.1
oauthlib==3.2.2
openai==1.59.8
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
packaging==24.2
pandas==2.2.3
pillow==10.4.0
portalocker==2.10.1
promptflow-core==1.17.1
promptflow-devkit==1.17.1
promptflow-tracing==1.17.1
protobuf==5.29.3
psutil==5.9.8
pycparser==2.22
pydantic==2.10.5
pydantic_core==2.27.2
pydash==7.0.7
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
ruamel.yaml==0.18.10
ruamel.yaml.clib==0.2.12
SecretStorage==3.3.3
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.41.3
strictyaml==1.7.3
tabulate==0.9.0
tiktoken==0.8.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
uvicorn==0.34.0
waitress==3.0.2
Werkzeug==3.1.3
wrapt==1.17.2
zipp==3.21.0
